Fixed non-working \donttest examples that came up when testing against R 3.2 devel per e-mail from BDR/CRAN 7 Dec
"""
Those parts of examples enclosed in \donttest{} are run by examples()
and so must be valid R code.  This is now checked by R-devel, and the
CRAN results pages for your packages show WARNINGs because some of your
code is invalid.

Please submit a corrected update checked with current R-devel (note that
the check stops at the first error: there may be more) as soon as
possible.

There may be other issues showing on the check pages which should also
be addressed.
"""