Display Region.Label as 'Region' in all dht output
Used to be 'Region' for summary table and 'Label' for N/D tables. Note this is only a change in how the dht object is printed and not how the values are stored so it won't break dependencies which may extract these values based on their object field names.  Closes #5